#26 #19 made methods final, added method.
Changed computeMedian to be final as well.
Changed mostUsedExercise to be final as well.

Created method to get the numbers for all the exercises used in a Pair
with Exercise and the number of times that Exercise was used, wrapped in
an ArrayList. The ArrayList is sorted such that the first Pair is the
one with highest frequency.