#20 #13
Added a writeObjects() in the removePerson(int id) method in
Database.java in main.app.core.data.
This resolves that the method removePerson actually updates the database
when it is called, such that persistence is maintained.
Earlier versions needed the method writeObjects() to be called
explicitly outside of removePerson(), which was counter-intuitive.

This was discovered when we tried to delete a person from the database
and then closing and restarting the app showed that the person still was
registered.